don't create training sets when model exists

+++ src/python/ranklib_learner.py
@@ -137,7 +137,10 @@ class RankLibLearner:
             test_ds = ranklib_ds[ranklib_ds["i_build"] == build]
             build_out_path = output_path / str(build)
             build_out_path.mkdir(parents=True, exist_ok=True)
-            if not (output_path / str(build) / "train.txt").exists():
+            if (
+                not (output_path / str(build) / "train.txt").exists()
+                and not (output_path / str(build) / "model.txt").exists()
+            ):
                 train_ds.to_csv(
                     output_path / str(build) / "train.txt",
                     sep=" ",
